+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.0 - 2022-09-15
+
+- mark as side-effect-free
+
 ## 1.2.0 - 2022-09-14
 
 - `divide()` function